refactor(resource): move to resource folder
