+++ domain/src/main/java/com/kaua/ecommerce/domain/resource/Resource.java
@@ -1,4 +1,4 @@
-package com.kaua.ecommerce.domain.utils;
+package com.kaua.ecommerce.domain.resource;
 
 import com.kaua.ecommerce.domain.ValueObject;
