+++ domain/src/test/java/com/kaua/ecommerce/domain/resource/ResourceTest.java
@@ -1,5 +1,7 @@
-package com.kaua.ecommerce.domain.utils;
+package com.kaua.ecommerce.domain.resource;
 
+import com.kaua.ecommerce.domain.resource.Resource;
+import com.kaua.ecommerce.domain.utils.RandomStringUtils;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
